Update UI to match GitHub's theme
Updated ui
