+++ PyLock/static/style.css
@@ -1,6 +1,6 @@
 body {
-    font-family: Arial, sans-serif;
-    background-color: #f4f4f4;
+    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif;
+    background-color: #f6f8fa;
     margin: 0;
     padding: 0;
 }
@@ -10,11 +10,12 @@ body {
     margin: 0 auto;
     padding: 20px;
     background-color: #fff;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+    box-shadow: 0 1px 3px rgba(27, 31, 35, 0.12), 0 8px 24px rgba(27, 31, 35, 0.12);
+    border-radius: 6px;
 }
 
 nav {
-    background-color: #333;
+    background-color: #24292e;
     color: #fff;
     padding: 10px;
 }
@@ -35,7 +36,7 @@ nav ul li a {
 }
 
 h1, h2 {
-    color: #333;
+    color: #24292e;
 }
 
 .form-group {
@@ -51,18 +52,21 @@ h1, h2 {
     width: 100%;
     padding: 8px;
     box-sizing: border-box;
+    border: 1px solid #d1d5da;
+    border-radius: 6px;
 }
 
 button {
     padding: 10px 15px;
-    background-color: #007bff;
+    background-color: #28a745;
     color: #fff;
     border: none;
     cursor: pointer;
+    border-radius: 6px;
 }
 
 button:hover {
-    background-color: #0056b3;
+    background-color: #218838;
 }
 
 #passwords-list, #search-results {
@@ -71,8 +75,9 @@ button:hover {
 
 #passwords-list div, #search-results div {
     padding: 10px;
-    border: 1px solid #ddd;
+    border: 1px solid #e1e4e8;
     margin-bottom: 10px;
+    border-radius: 6px;
 }
 
 /* Additional styles for password management UI */
