+++ project/templates/upload.html
@@ -1,41 +1,3 @@
-<!--
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Upload Excel File</title>
-</head>
-<body>
-    <h1>Upload Policy Details</h1>
-    <form action="/" method="post" enctype="multipart/form-data">
-        <div>
-            <label for="file">Upload Excel File:</label>
-            <input type="file" name="file" accept=".xlsx">
-        </div>
-        <div>
-            <h2>Or Enter Policy Details Manually</h2>
-            <label for="policy_name">Policy Name:</label>
-            <input type="text" name="policy_name" required>
-        </div>
-        <div>
-            <label for="policy_end_date">Policy End Date:</label>
-            <input type="date" name="policy_end_date" required>
-        </div>
-        <div>
-            <label for="notify_before">Notify Before:</label>
-            <select name="notify_before" required>
-                <option value="1">1 day</option>
-                <option value="2">2 days</option>
-                <option value="5">5 days</option>
-                <option value="10">10 days</option>
-            </select>
-        </div>
-        <button type="submit">Submit</button>
-    </form>
-</body>
-</html>
--->
-
 {% extends "base.html" %}
 
 {% block content %}
